Fix undo/redo for filters
