+++ handsontable/src/plugins/filters/__tests__/filters.spec.js
@@ -657,147 +657,6 @@ describe('Filters', () => {
     });
   });
 
-  describe('Undo/Redo', () => {
-    it('should undo previously added filters', () => {
-      const hot = handsontable({
-        data: getDataForFilters(),
-        columns: getColumnsForFilters(),
-        dropdownMenu: true,
-        filters: true,
-        width: 500,
-        height: 300
-      });
-      const plugin = hot.getPlugin('filters');
-
-      plugin.addCondition(0, 'gt', [3]);
-      plugin.filter();
-      plugin.addCondition(2, 'begins_with', ['b']);
-      plugin.filter();
-      plugin.addCondition(4, 'eq', ['green']);
-      plugin.filter();
-
-      expect(getData().length).toEqual(2);
-
-      getPlugin('undoRedo').undo();
-
-      expect(getData().length).toEqual(3);
-
-      getPlugin('undoRedo').undo();
-
-      expect(getData().length).toEqual(36);
-
-      getPlugin('undoRedo').undo();
-
-      expect(getData().length).toEqual(39);
-    });
-
-    it('should redo previously reverted filters', () => {
-      const hot = handsontable({
-        data: getDataForFilters(),
-        columns: getColumnsForFilters(),
-        dropdownMenu: true,
-        filters: true,
-        width: 500,
-        height: 300
-      });
-      const plugin = hot.getPlugin('filters');
-
-      plugin.addCondition(0, 'gt', [3]);
-      plugin.filter();
-      plugin.addCondition(2, 'begins_with', ['b']);
-      plugin.filter();
-      plugin.addCondition(4, 'eq', ['green']);
-      plugin.filter();
-
-      getPlugin('undoRedo').undo();
-      getPlugin('undoRedo').undo();
-      getPlugin('undoRedo').undo();
-
-      expect(getData().length).toEqual(39);
-
-      getPlugin('undoRedo').redo();
-
-      expect(getData().length).toEqual(36);
-
-      getPlugin('undoRedo').redo();
-
-      expect(getData().length).toEqual(3);
-
-      getPlugin('undoRedo').redo();
-
-      expect(getData().length).toEqual(2);
-    });
-
-    it('should undo multiple steps of filtering performed with the Filters\' UI', async() => {
-      handsontable({
-        data: getDataForFilters(),
-        columns: getColumnsForFilters(),
-        dropdownMenu: true,
-        filters: true,
-        width: 500,
-        height: 300
-      });
-
-      dropdownMenu(1);
-      openDropdownByConditionMenu();
-      selectDropdownByConditionMenuOption('Begins with');
-
-      await sleep(20);
-      document.activeElement.value = 'R';
-      keyUp('R');
-
-      $(dropdownMenuRootElement().querySelector('.htUIButton.htUIButtonOK input')).simulate('click');
-
-      expect(getData().length).toEqual(2);
-
-      setDataAtCell(0, 1, `${getDataAtCell(0, 1)}!`);
-      setDataAtCell(1, 1, `${getDataAtCell(1, 1)}!`);
-
-      expect(getDataAtCell(0, 1).includes('!')).toBe(true);
-      expect(getDataAtCell(1, 1).includes('!')).toBe(true);
-
-      dropdownMenu(1);
-      openDropdownByConditionMenu();
-      selectDropdownByConditionMenuOption('Begins with');
-
-      await sleep(20);
-      document.activeElement.value = '';
-      keyUp('Backspace');
-
-      $(dropdownMenuRootElement().querySelector('.htUIButton.htUIButtonOK input')).simulate('click');
-
-      expect(getData().length).toEqual(39);
-      expect(getDataAtCell(0, 1).includes('!')).toBe(false);
-      expect(getDataAtCell(1, 1).includes('!')).toBe(false);
-      expect(getDataAtCell(4, 1).includes('!')).toBe(true);
-      expect(getDataAtCell(33, 1).includes('!')).toBe(true);
-
-      getPlugin('undoRedo').undo();
-
-      expect(getData().length).toEqual(2);
-      expect(getDataAtCell(0, 1).includes('!')).toBe(true);
-      expect(getDataAtCell(1, 1).includes('!')).toBe(true);
-
-      getPlugin('undoRedo').undo();
-
-      expect(getDataAtCell(0, 1).includes('!')).toBe(true);
-      expect(getDataAtCell(1, 1).includes('!')).toBe(false);
-
-      getPlugin('undoRedo').undo();
-
-      expect(getDataAtCell(0, 1).includes('!')).toBe(false);
-      expect(getDataAtCell(1, 1).includes('!')).toBe(false);
-
-      getPlugin('undoRedo').undo();
-
-      expect(getData().length).toEqual(39);
-      expect(getDataAtCell(0, 1).includes('!')).toBe(false);
-      expect(getDataAtCell(1, 1).includes('!')).toBe(false);
-      expect(getDataAtCell(4, 1).includes('!')).toBe(false);
-      expect(getDataAtCell(33, 1).includes('!')).toBe(false);
-    });
-  });
-
   describe('Hooks', () => {
     describe('`beforeFilter` hook', () => {
       it('should trigger `beforeFilter` hook after filtering values', () => {
